Rename to CEREsFit
Fix lint
