+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2022 Galactic Forensics Laboratory
+Copyright (c) 2023 Reto Trappitsch
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
